// src/app/cyprus/types.ts
export interface CyprusFlight {
  flightHashData: string
  currency: string
  flightType: string
  flightSegmentList: {
    origin: string
    destination: string
    flightType: string | null
    flightList: {
      departureDate: string
      departureTime: string
      arrivalDate: string
      arrivalTime: string
      flightDetails: {
        flightNo: string
        airline: string
        flightSource: string
        departureDate: string
        departureTime: string
        arrivalDate: string
        arrivalTime: string
        origin: string
        destination: string
        diffPrice?: ServicePriceType
        voyageCode: null
        className: string
        classType: string
        seat: number
        totalPrice: ServicePriceType
        key: null
        totalTax: ServicePriceType
        totalServiceFee: ServicePriceType
        priceCode: string
        basePrice: ServicePriceType
        taxes: ServicePriceType
        discount: ServicePriceType
        buyFee: {
          code: null
          price: ServicePriceType
        }
        fee: {
          code: null
          price: ServicePriceType
        }
        passengerPrices: null
        taxInfos: null
        serviceCharges: null
      }[]
    }[]
  }[]
}